Split out leaderboard viewport
