+++ internal/resources/Leaderboard.go
@@ -9,6 +9,8 @@ import (
 
 	"dalton.dog/aocgo/internal/api"
 	"dalton.dog/aocgo/internal/cache"
+	"dalton.dog/aocgo/internal/tui"
+
 	// "dalton.dog/aocgo/internal/tui"
 	"github.com/PuerkitoBio/goquery"
 	"github.com/charmbracelet/lipgloss"
@@ -104,7 +106,7 @@ func (l *Leaderboard) Display() {
 		t.Row(strconv.Itoa(p.placement), strconv.Itoa(p.score), p.displayName)
 	}
 
-	// tui.StartViewportWithString(t.Render(), fmt.Sprintf("Leaderboard - %d", l.year), false)
+	tui.NewLeaderboardViewport(t.Render(), fmt.Sprintf("Leaderboard - %d", l.year))
 }
 
 func (l *Leaderboard) loadPlacements() error {
