--- internal/tui/viewport.go
@@ -1,1 +1,0 @@
-package tui
